Fix CodeQL missing rate limiting on analytics server smoke test.
Extract configureApp into analytics-app.js so tests use the same rate-limited /collect route as production.
