+++ server/server-stack.test.js
@@ -2,6 +2,7 @@
 
 const { describe, it, before, after } = require('node:test');
 const assert = require('node:assert/strict');
+const { configureApp } = require('./analytics-app');
 
 /** @param {import('node:http').Server} server */
 const getServerPort = (server) => {
@@ -14,36 +15,18 @@ const getServerPort = (server) => {
 
 describe('analytics server stack (express 5)', () => {
   /** @type {import('node:http').Server | undefined} */
-  let server = undefined;
+  let server;
 
   before(async () => {
     const express = require('express');
     const fs = require('node:fs');
     const os = require('node:os');
     const path = require('node:path');
-    const { registerBillingRoutes } = require('./billing');
-    const { registerBetaRoutes } = require('./beta');
-    const { registerMetricsRoutes } = require('./metrics');
-    const { registerSentryTunnelRoutes } = require('./sentry-tunnel');
-    const rateLimit = require('express-rate-limit');
 
     const logFile = path.join(os.tmpdir(), `gfc-analytics-test-${process.pid}.log`);
 
     const app = express();
-    app.set('trust proxy', 'loopback');
-
-    registerSentryTunnelRoutes(app, express, rateLimit);
-    registerBillingRoutes(app, express);
-    registerMetricsRoutes(app, express);
-    registerBetaRoutes(app, express);
-
-    app.post('/collect', express.json({ limit: '1kb' }), (req, res) => {
-      const page = (typeof req.body?.page === 'string' ? req.body.page : '/').slice(0, 200);
-      fs.appendFileSync(logFile, `${JSON.stringify({ page })}\n`);
-      res.status(204).end();
-    });
-
-    app.use((_req, res) => res.status(404).end());
+    configureApp(app, express, fs, logFile);
 
     server = await new Promise((resolve, reject) => {
       const instance = app.listen(0, '127.0.0.1', () => resolve(instance));
